backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.0
PyPDF2>=3.0.1
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.9.0
sentence-transformers>=3.0.0
groq>=0.9.0
python-dotenv>=1.0.1